Refactor: useCountdown으로 리팩토링
